requirements.txt
# Tested on Python 3.11
numpy==1.26.4
scipy==1.17.1
pandas==2.2.2
scikit-learn==1.4.2
statsmodels==0.14.2
matplotlib==3.8.4
tqdm==4.66.4
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pytest==8.2.2
wrds==3.5.0
python-glmnet==2.6.1
pytz 
python-dateutil
tzdata 
packaging 
patsy 
contourpy 
cycler 
fonttools 
kiwisolver 
pillow 
pyparsing
iniconfig 
pluggy
joblib
pyarrow>=15.0
jinja2>=3.1

